dm-paddle/requirements.txt
paddlepaddle == 2.3.2
matplotlib == 3.6.3